\name{Power}
\alias{Power}
\title{Extract the Power results}
\description{
  Extract the Power results from a compound object into a Stacked Dataframe
}
\usage{
Power(object, subset)
}
\arguments{
  \item{object}{an object of class \code{cpd.PwrGSD}}
  \item{subset}{you may extract a subset via a logical expression in the
    variables of the index dataframe, \code{descr}}
}
\value{
  an object of class \code{cpd.PwrGSD}. See help on that topic for details.
}
\author{Grant Izmirlian <izmirlian@nih.gov>}
\seealso{\code{\link{cpd.PwrGSD}} and \code{\link{PwrGSD}}}
\examples{
  ## See the `cpd.PwrGSD' example
}
\keyword{data}
