\name{DX}
\alias{DX}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{A utility function for forming differences}
\description{
  DX(x) returns c(x[1], diff(x))
}
\usage{
  DX(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A grid of time points (increasing)}
}
\value{
  DX(x) returns c(x[1], diff(x))
}
\author{Grant Izmirlian <izmirlian@nih.gov}
\keyword{misc}
