\name{RCM2RR}
\alias{RCM2RR}
\title{Relative cumulative mortality to Relative Risk}
\description{
  Relative cumulative mortality to Relative Risk
}
\usage{
  RCM2RR(tlook, tcut.i, h.i, hOth, accru, rcm)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tlook}{}
  \item{tcut.i}{}
  \item{h.i}{}
  \item{hOth}{}
  \item{accru}{}
  \item{rcm}{}
}

\value{
  ~Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
}
\author{Grant Izmirlian <izmirlian@nih.gov>}
\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{
 ## none as yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{survival}
